add captcha

+++ backend/src/main/java/com/eshop/backend/product/catalog/service/MainService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface MainService<T> {
     void create(T model);
-    T getById(Long id);
+    T getById(int id);
     List<T> getAll();
     void update(T model);
     void delete(Long id);
